% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_legend_text.R
\name{change_legend_text}
\alias{change_legend_text}
\title{change_legend_text()}
\usage{
change_legend_text(
  svg_xml,
  orig_text,
  change_text = NULL,
  text_color = "black",
  save_as = NULL
)
}
\arguments{
\item{svg_xml}{Your XML document with legend.}

\item{orig_text}{Original text in the legend.}

\item{change_text}{Text you want to replace.}

\item{text_color}{The color of text.}

\item{save_as}{A file path to save the modified SVG. Default is NULL.}
}
\value{
Changed SVG file.
}
\description{
This function help you change the text and character of text in the legend.
}
