% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{find_path_id}
\alias{find_path_id}
\title{find_path_id()
This function finds the IDs of \if{html}{\out{<path>}} elements in an SVG file.}
\usage{
find_path_id(svg_xml)
}
\arguments{
\item{svg_xml}{XML document from read_n_check(infile).}
}
\value{
A vector of path IDs (printed and returned).
}
\description{
find_path_id()
This function finds the IDs of \if{html}{\out{<path>}} elements in an SVG file.
}
