% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmapSVG.R
\name{heatmapSVG}
\alias{heatmapSVG}
\title{heatmapSVG()
This function helps you color path with gradation based on your data.}
\usage{
heatmapSVG(
  data_mapped,
  svg_xml,
  data_col,
  id_col = "matched_path_id",
  range,
  palette,
  breaks = NULL,
  direction = "vertical",
  position = NULL,
  size = NULL,
  label_num = 3,
  save_as = NULL
)
}
\arguments{
\item{data_mapped}{A dataframe containing the mapping results from mapping_path_id.}

\item{svg_xml}{XML document from read_n_check(infile).}

\item{data_col}{Column name of value that is used to make heatmap.}

\item{id_col}{Column name of id. Default matched_path_id,}

\item{range}{Under and upper limit of the value. You should give this value since it determines the color. Format should be vector; c(0, 1).}

\item{palette}{Color you want. Format should be vector; c("blue", "red").}

\item{breaks}{Breaks for discrete heatmap. If provided, a discrete heatmap will be generated.}

\item{direction}{Default is vertical. You can choose between vertical and horizontal.}

\item{position}{Default is right middle (0.9 of width, 0.33 of height). Format should be vector, for example c(0.9, 0.33).}

\item{size}{Default is 0.03 of width, 0.33 of height. Format should be vector, for example c(0.03, 0.33)}

\item{label_num}{Default is 3, it designate how much numbers would show next to colorbar.}

\item{save_as}{A file path to save the modified SVG. Default is NULL.}
}
\value{
Modified XML document with updated colors (print and return).
}
\description{
heatmapSVG()
This function helps you color path with gradation based on your data.
}
