% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mapping_path_id}
\alias{mapping_path_id}
\title{mapping_path_id()
This function maps the path IDs from an SVG file to the corresponding labels in a user-provided dataframe.}
\usage{
mapping_path_id(data, path_id, col_name, col_factor)
}
\arguments{
\item{data}{A dataframe containing the labeled regions.}

\item{path_id}{A vector of path IDs from the SVG file (obtained from find_path_id()).}

\item{col_name}{The name of the column in \code{data} that contains the region labels.}

\item{col_factor}{A character vector specifying the order of labels for matching with \code{path_id}.}
}
\value{
The original dataframe with an additional column containing matched path IDs.
}
\description{
The function takes a dataframe with labeled regions and a vector of path IDs from an SVG file.
It maps these IDs to the dataframe based on a specified order in \code{col_factor} and adds a new column with matched IDs.
}
