% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_n_check.R
\name{read_n_check}
\alias{read_n_check}
\title{read_n_check()
This function reads an SVG file as an XML document and displays it as a raster plot for checking.}
\usage{
read_n_check(infile)
}
\arguments{
\item{infile}{Path to the input SVG file. The format should be "path/to/your/infile"}
}
\value{
None (just displays the plot)
}
\description{
read_n_check()
This function reads an SVG file as an XML document and displays it as a raster plot for checking.
}
