% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{recolorSVG}
\alias{recolorSVG}
\title{recolorSVG
This function changes the colors of an SVG file by modifying the "fill" property in the style attribute of path elements.
Use this function when you want to apply specific colors to specific paths.}
\usage{
recolorSVG(svg_xml, path_id, color_factor, save_as = NULL)
}
\arguments{
\item{svg_xml}{An XML document from read_n_check(infile).}

\item{path_id}{A character vector of path IDs to be colored.}

\item{color_factor}{A character vector of colors. The order should match the order of \code{path_id}.}

\item{save_as}{A file path to save the modified SVG. Default is NULL.}
}
\value{
Modified XML document with updated colors (print and return).
}
\description{
recolorSVG
This function changes the colors of an SVG file by modifying the "fill" property in the style attribute of path elements.
Use this function when you want to apply specific colors to specific paths.
}
