% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_guideline.R
\name{xy_guideline}
\alias{xy_guideline}
\title{xy_guideline()}
\usage{
xy_guideline(svg_xml, color = "red")
}
\arguments{
\item{svg_xml}{File XML document from read_n_check(infile).}

\item{color}{Color of the guideline. Default is red.}
}
\value{
Rastered image with 10*10 guidelines
}
\description{
This function help you how to set your position and size of the legend.
}
